requirements.txt
Django==4.1.5
pyTelegramBotAPI==4.9.0
djangorestframework==3.14.0
djoser==2.1.0